work in progress on retention
