+++ internal/controller/translator/v2/retention.go
@@ -1,11 +1,26 @@
 package v2
 
-import wandbv2 "github.com/wandb/operator/api/v2"
+import (
+	wandbv2 "github.com/wandb/operator/api/v2"
+	"k8s.io/apimachinery/pkg/labels"
+)
 import "github.com/wandb/operator/internal/controller/translator"
 
-func ToOnDeletePolicy(retentionPolicy wandbv2.WBRetentionPolicy) translator.OnDeletePolicy {
+func ToOnDeleteRule(wandb *wandbv2.WeightsAndBiases, retentionPolicy wandbv2.WBRetentionPolicy) translator.OnDeleteRule {
+	policy := translator.Preserve
 	if retentionPolicy.OnDelete == wandbv2.WBPurgeOnDelete {
-		return translator.Purge
+		policy = translator.Purge
 	}
-	return translator.Preserve
+	selector := labels.SelectorFromSet(retentionLabelMap(wandb))
+	return translator.OnDeleteRule{
+		Policy:   policy,
+		Selector: selector,
+	}
+}
+
+func retentionLabelMap(wandb *wandbv2.WeightsAndBiases) map[string]string {
+	return map[string]string{
+		translator.WandbNameLabel:      wandb.Name,
+		translator.WandbNamespaceLabel: wandb.Namespace,
+	}
 }
